refactor(negotiation): move exceptions to errors
