+++ src/a2a_t/negotiation/common/errors.py
