+++ src/a2a_t/negotiation/common/models.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from .exceptions import NegotiationContextError
+from .errors import NegotiationContextError
 from .enums import NegotiationRole, NegotiationStatus, NegotiationType
